--- extension-manifest-v3/src/content_scripts/iframe.js
@@ -1,26 +1,0 @@
-/**
- * Ghostery Browser Extension
- * https://www.ghostery.com/
- *
- * Copyright 2017-present Ghostery GmbH. All rights reserved.
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0
- */
-
-import { showIframe } from '@ghostery/ui/iframe';
-
-let shown = false;
-chrome.runtime.onMessage.addListener((msg) => {
-  if (msg.action === 'showIframe') {
-    if (shown) return false;
-
-    showIframe(chrome.runtime.getURL(msg.url));
-    shown = true;
-
-    return false;
-  }
-
-  return false;
-});
