+++ extension-manifest-v3/src/content_scripts/youtube.js
@@ -9,8 +9,48 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0
  */
 
-import { showIframe, closeIframe } from '@ghostery/ui/iframe';
-import detectWall from '@ghostery/ui/youtube/wall';
+import * as notifications from '/utils/notifications.js';
+
+const SELECTORS = [
+  // Based on https://github.com/AdguardTeam/AdguardFilters/blob/e5ae8e3194f8d18bdcc660d4c42282e4a96ca5b9/AnnoyancesFilter/Popups/sections/antiadblock.txt#L2044
+  'ytd-watch-flexy:not([hidden]) ytd-enforcement-message-view-model > div.ytd-enforcement-message-view-model',
+
+  'yt-playability-error-supported-renderers#error-screen ytd-enforcement-message-view-model',
+  'tp-yt-paper-dialog .ytd-enforcement-message-view-model',
+];
+
+function detectWall(cb) {
+  let timeout = null;
+
+  const observer = new MutationObserver(() => {
+    if (timeout) return;
+
+    timeout = setTimeout(() => {
+      if (document.querySelector(SELECTORS)?.clientHeight > 0) {
+        try {
+          cb();
+        } catch (e) {
+          /* ignore */
+        }
+      } else {
+        timeout = null;
+      }
+    }, 1000 /* 1 second delay */);
+  });
+
+  document.addEventListener('yt-navigate-start', () => {
+    clearTimeout(timeout);
+    timeout = null;
+  });
+
+  document.addEventListener('DOMContentLoaded', () => {
+    observer.observe(document.body, {
+      childList: true,
+      subtree: true,
+      attributeFilter: ['src', 'style'],
+    });
+  });
+}
 
 async function isFeatureDisabled() {
   const { options, youtubeDontAsk } = await chrome.storage.local.get([
@@ -40,19 +80,21 @@ if (!chrome.extension.inIncognitoContext) {
   (async () => {
     if (await isFeatureDisabled()) return;
 
-    window.addEventListener('yt-navigate-start', () => closeIframe(), true);
-
     detectWall(async () => {
       if (await isFeatureDisabled()) return;
-
-      showIframe(
-        chrome.runtime.getURL(
-          `/pages/youtube/index.html?url=${encodeURIComponent(
-            window.location.href,
-          )}`,
-        ),
-        '460px',
-      );
+      chrome.runtime.sendMessage({
+        action: notifications.OPEN_ACTION,
+        id: 'youtube',
+        params: { url: window.location.href },
+      });
     });
+
+    window.addEventListener(
+      'yt-navigate-start',
+      () => {
+        chrome.runtime.sendMessage({ action: notifications.CLOSE_ACTION });
+      },
+      true,
+    );
   })();
 }
