+++ extension-manifest-v3/src/manifest.firefox.json
@@ -60,7 +60,7 @@
     {
       "matches": ["http://*/*", "https://*/*"],
       "run_at": "document_start",
-      "js": ["content_scripts/iframe.js"]
+      "js": ["content_scripts/notifications.js"]
     },
     {
       "matches": ["http://*/*", "https://*/*"],
@@ -78,7 +78,7 @@
     "content_scripts/prevent-serp-tracking.js",
     "pages/trackers-preview/index.html",
     "pages/onboarding/index.html",
-    "pages/youtube/index.html"
+    "pages/notifications/youtube.html"
   ],
   "content_security_policy": "script-src 'self'",
   "browser_specific_settings": {
