+++ extension-manifest-v3/src/manifest.safari-macos.json
@@ -91,7 +91,7 @@
       "matches": ["http://*/*", "https://*/*"],
       "run_at": "document_start",
       "js": [
-        "content_scripts/iframe.js"
+        "content_scripts/notifications.js"
       ]
     },
     {
@@ -113,7 +113,7 @@
     "content_scripts/prevent-serp-tracking.js",
     "pages/trackers-preview/index.html",
     "pages/onboarding/index.html",
-    "pages/youtube/index.html"
+    "pages/notifications/youtube.html"
   ],
   "content_security_policy": "script-src 'self'",
   "browser_specific_settings": {
