+++ extension-manifest-v3/src/pages/notifications/youtube.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width" />
     <title>Youtube Wall</title>
-    <script src="./index.js" type="module" async></script>
+    <script src="./youtube.js" type="module" async></script>
   </head>
   <body></body>
 </html>
