+++ packages/ui/package.json
@@ -7,12 +7,9 @@
   "exports": {
     "./onboarding": "./src/modules/onboarding/index.js",
     "./panel": "./src/modules/panel/index.js",
-    "./youtube": "./src/modules/youtube/index.js",
-    "./youtube/wall": "./src/modules/youtube/wall.js",
     "./categories": "./src/utils/categories.js",
     "./labels": "./src/utils/labels.js",
-    "./wheel": "./src/utils/wheel.js",
-    "./iframe": "./src/utils/iframe.js"
+    "./wheel": "./src/utils/wheel.js"
   },
   "scripts": {
     "lint": "eslint ./src",
